Now color function takes a float instea of an int

+++ FractalVisualizer/assets/colors/3cycle.glsl
@@ -1,11 +1,11 @@
-#uniform float colorMult 100 1 3 NULL false;
+#uniform float colorMult 100 1 1 NULL false;
 #uniform float offset 0 0.01 -1 1;
 #uniform color c1 0.20 0.10 0.07;
 #uniform color c2 0.14 0.39 0.45;
 #uniform color c3 0.96 0.70 0.18;
 
 vec3 colors[] = vec3[](c1, c2, c3, c1);
-vec3 get_color(int iters)
+vec3 get_color(float iters)
 {
 	float x = (iters / colorMult + offset) * (colors.length() - 1);
     x = mod(x, colors.length() - 1);
