+++ FractalVisualizer/assets/colors/4cycle.glsl
@@ -1,4 +1,4 @@
-#uniform float colorMult 100 1 4 NULL false;
+#uniform float colorMult 100 1 1 NULL false;
 #uniform float offset 0 0.01 -1 1;
 #uniform color c1 0.00 0.00 0.00;
 #uniform color c2 0.60 0.22 0.26;
@@ -6,7 +6,7 @@
 #uniform color c4 0.13 0.31 0.50;
 
 vec3 colors[] = vec3[](c1, c2, c3, c4, c1);
-vec3 get_color(int iters)
+vec3 get_color(float iters)
 {
 	float x = (iters / colorMult + offset) * (colors.length() - 1);
     x = mod(x, colors.length() - 1);
