+++ FractalVisualizer/assets/colors/trans.glsl
@@ -1,7 +1,7 @@
 #uniform float colorMult 100 1 1 NULL false;
 #uniform float offset 0 0.02 -1 1;
 #define PI 3.1415926535
-vec3 get_color(int i)
+vec3 get_color(float i)
 {
     float x = i / colorMult + offset;
     float n1 = sin(2*PI*x) * 0.5 + 0.5;
